refactor(enums): change enums package
