+++ src/main/java/com/jiuzhang/flashsale/component/RedisPreheatRunner.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import javax.annotation.Resource;
 
+import com.jiuzhang.flashsale.common.enums.ActivityStatus;
 import com.jiuzhang.flashsale.entity.ActivityEntity;
-import com.jiuzhang.flashsale.enums.ActivityStatus;
 import com.jiuzhang.flashsale.exception.RedisStockException;
 import com.jiuzhang.flashsale.service.ActivityService;
 import com.jiuzhang.flashsale.service.impl.RedisServiceImpl;
