+++ src/main/java/com/jiuzhang/flashsale/entity/ActivityEntity.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
-import com.jiuzhang.flashsale.enums.ActivityStatus;
+import com.jiuzhang.flashsale.common.enums.ActivityStatus;
 
 import org.springframework.format.annotation.DateTimeFormat;
