+++ src/main/java/com/jiuzhang/flashsale/entity/OrderEntity.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-import com.jiuzhang.flashsale.enums.OrderStatus;
+import com.jiuzhang.flashsale.common.enums.OrderStatus;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
