+++ src/main/java/com/jiuzhang/flashsale/mq/PayStatusCheckConsumer.java
@@ -5,8 +5,8 @@ import java.nio.charset.StandardCharsets;
 import javax.annotation.Resource;
 
 import com.alibaba.fastjson.JSON;
+import com.jiuzhang.flashsale.common.enums.OrderStatus;
 import com.jiuzhang.flashsale.entity.OrderEntity;
-import com.jiuzhang.flashsale.enums.OrderStatus;
 import com.jiuzhang.flashsale.exception.RedisStockException;
 import com.jiuzhang.flashsale.exception.RedisUserException;
 import com.jiuzhang.flashsale.service.ActivityService;
