+++ src/main/java/com/jiuzhang/flashsale/service/ActivityService.java
@@ -3,8 +3,8 @@ package com.jiuzhang.flashsale.service;
 import java.util.List;
 
 import com.baomidou.mybatisplus.extension.service.IService;
+import com.jiuzhang.flashsale.common.enums.ActivityStatus;
 import com.jiuzhang.flashsale.entity.ActivityEntity;
-import com.jiuzhang.flashsale.enums.ActivityStatus;
 import com.jiuzhang.flashsale.exception.RedisStockException;
 
 /**
