+++ src/test/java/com/jiuzhang/flashsale/mapper/ActivityMapperTest.java
@@ -7,8 +7,8 @@ import java.time.LocalDateTime;
 
 import javax.annotation.Resource;
 
+import com.jiuzhang.flashsale.common.enums.ActivityStatus;
 import com.jiuzhang.flashsale.entity.ActivityEntity;
-import com.jiuzhang.flashsale.enums.ActivityStatus;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
